残差能量打印正确，但1024乘1024的图像，边界填充值为128的残差能量为153501690

+++ intra_lss/pre.h
@@ -53,7 +53,7 @@ static void predict_4x4_hu(uint8_t *src);
 	const int t6 = src[6-1*FDEC_STRIDE];\
 	const int t7 = src[7-1*FDEC_STRIDE]; 
 
-short predict(short image_construct[1025][1029], short resi[1024][1024], short predicted[1024][1024], int height, int width);
+long long predict(short image_construct[1025][1029], short resi[1024][1024], short predicted[1024][1024], int height, int width);
 void DCTCore4x4(uint8_t block[4][4], uint8_t tblock[4][4]);
 void IDCTCore4x4(uint8_t tblock[4][4], uint8_t block[4][4]);
